@@ -12,21 +12,18 @@
         }
     ],
     "require": {
-        "php": "^7.2 || ^8",
-        "symfony/framework-bundle": "^4.3|^5.0",
-        "symfony/yaml": "^4.3|^5.0",
+        "php": "^8",
+        "symfony/framework-bundle": "^5.1",
+        "symfony/yaml": "^5.1",
         "mmoreram/symfony-bundle-dependencies": "^2.3"
     },
     "require-dev": {
-        "symfony/http-kernel": "^4.3|^5.0",
-        "symfony/console": "^4.3|^5.0",
-        "symfony/process":  "^4.3|^5.0",
-        "symfony/config": "^4.3|^5.0",
-        "symfony/browser-kit": "^4.3|^5.0",
-        "phpunit/phpunit": ">=8.0"
-    },
-    "conflict": {
-        "phpunit/phpunit": "<8"
+        "symfony/http-kernel": "^5.1",
+        "symfony/console": "^5.1",
+        "symfony/process":  "^5.1",
+        "symfony/config": "^5.1",
+        "symfony/browser-kit": "^5.1",
+        "phpunit/phpunit": "^9"
     },
     "autoload": {
         "psr-4": {
